Create basic layout and dashboard content
Layout is as intended, while some of the padding and margins
for each section must be corrected for the stretched out version.
Narrow version is now working correctly, but some responsiveness
will be necessary for correct display on full screen mode.
Links and icons are styled in a basic but correct way, with room
to improve.
Also needs svg icons and images added to complete the basic design.